+++ src/data/layout/array/carrier/mod.rs
@@ -1,0 +1,12 @@
+// devela/src/data/layout/array/carrier/mod.rs
+//
+//!
+//
+
+mod slice;
+// mod fixed;
+
+// #[cfg(feature = "alloc")]
+// mod boxed;
+// #[cfg(feature = "alloc")]
+// mod vec;