+++ src/data/layout/queue/destaque/define.rs
@@ -36,7 +36,7 @@ use crate::{Array, Bare, Storage};
 /// [`DestaqueU8`], [`DestaqueU16`], [`DestaqueU32`], [`DestaqueUsize`],
 /// and the related traits:
 /// [`DataQueue`][crate::DataQueue], [`DataDeque`][crate::DataDeque],
-/// [`DataStack`][crate::DataStack], [`DataDesta`][crate::DataDesta].<br/>
+/// [`DataStack`][crate::DataStack].<br/>
 ///
 /// ## Methods
 ///